icashx-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.5.3
python-dotenv==1.0.1
requests==2.32.3
openai==1.12.0
twilio==9.2.1
python-docx==1.1.0
psycopg2-binary==2.9.9
celery==5.4.0
python-multipart==0.0.9
jinja2==3.1.3